+++ rumqttc/src/v5/framed.rs
@@ -1,12 +1,14 @@
-use futures_util::{FutureExt, SinkExt};
+use bytes::Buf;
+use futures_util::SinkExt;
 use tokio_stream::StreamExt;
 use tokio_util::codec::Framed;
 
 use crate::framed::AsyncReadWrite;
 
-use super::mqttbytes::v5::Packet;
-use super::{mqttbytes, Codec, Connect, MqttOptions, MqttState};
-use super::{Incoming, StateError};
+use super::{
+    mqttbytes::v5::Packet,
+    Codec, {Incoming, StateError},
+};
 
 /// Network transforms packets <-> frames efficiently. It takes
 /// advantage of pre-allocation, buffering and vectorization when
@@ -14,8 +16,6 @@ use super::{Incoming, StateError};
 pub struct Network {
     /// Frame MQTT packets from network connection
     framed: Framed<Box<dyn AsyncReadWrite>, Codec>,
-    /// Maximum readv count
-    max_readb_count: usize,
 }
 impl Network {
     pub fn new(socket: impl AsyncReadWrite + 'static, max_incoming_size: Option<u32>) -> Network {
@@ -26,56 +26,24 @@ impl Network {
         };
         let framed = Framed::new(socket, codec);
 
-        Network {
-            framed,
-            max_readb_count: 10,
-        }
+        Network { framed }
     }
 
     pub fn set_max_outgoing_size(&mut self, max_outgoing_size: Option<u32>) {
         self.framed.codec_mut().max_outgoing_size = max_outgoing_size;
     }
 
-    /// Reads and returns a single packet from network
-    pub async fn read(&mut self) -> Result<Incoming, StateError> {
-        match self.framed.next().await {
-            Some(Ok(packet)) => Ok(packet),
-            Some(Err(mqttbytes::Error::InsufficientBytes(_))) => unreachable!(),
-            Some(Err(e)) => Err(StateError::Deserialization(e)),
-            None => Err(StateError::ConnectionAborted),
-        }
+    pub fn read_buffer_remaining(&self) -> usize {
+        self.framed.read_buffer().remaining()
     }
 
-    /// Read packets in bulk. This allow replies to be in bulk. This method is used
-    /// after the connection is established to read a bunch of incoming packets
-    pub async fn readb(&mut self, state: &mut MqttState) -> Result<(), StateError> {
-        // wait for the first read
-        let mut res = self.framed.next().await;
-        let mut count = 1;
-        loop {
-            match res {
-                Some(Ok(packet)) => {
-                    if let Some(outgoing) = state.handle_incoming_packet(packet)? {
-                        self.write(outgoing).await?;
-                    }
-
-                    count += 1;
-                    if count >= self.max_readb_count {
-                        break;
-                    }
-                }
-                Some(Err(mqttbytes::Error::InsufficientBytes(_))) => unreachable!(),
-                Some(Err(e)) => return Err(StateError::Deserialization(e)),
-                None => return Err(StateError::ConnectionAborted),
-            }
-            // do not wait for subsequent reads
-            match self.framed.next().now_or_never() {
-                Some(r) => res = r,
-                _ => break,
-            };
+    /// Reads and returns a single packet from network
+    pub async fn read(&mut self) -> Result<Option<Incoming>, StateError> {
+        match self.framed.next().await {
+            Some(Ok(packet)) => Ok(Some(packet)),
+            Some(Err(e)) => Err(StateError::Deserialization(e)),
+            None => Ok(None),
         }
-
-        Ok(())
     }
 
     /// Serializes packet into write buffer
@@ -86,19 +54,7 @@ impl Network {
             .map_err(StateError::Deserialization)
     }
 
-    pub async fn connect(
-        &mut self,
-        connect: Connect,
-        options: &MqttOptions,
-    ) -> Result<(), StateError> {
-        let last_will = options.last_will();
-        let login = options.credentials();
-        self.write(Packet::Connect(connect, last_will, login))
-            .await?;
-
-        self.flush().await
-    }
-
+    /// Flush the outgoing sink
     pub async fn flush(&mut self) -> Result<(), StateError> {
         self.framed
             .flush()